Adv-Yolo-master/requirements.txt
absl-py ==2.3.0
accelerate ==1.5.2
aiofiles ==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
asttokens ==3.0.0
attrs==25.3.0
bitsandbytes==0.46.1
blobfile==3.0.0
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cmake==3.18.4
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
datasets==3.4.1
debugpy==1.8.11
decorator==5.2.1
deepspeed==0.16.5
dill==0.3.8
distro==1.9.0
dotenv==0.9.9
editdistance==0.8.1
einops==0.8.1
evaluate==0.4.3
executing==2.1.0
fastapi==0.115.12
ffmpy==0.5.0
filelock==3.13.1
fire==0.7.0
flash_attn==2.7.4.post1
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.6.1
gradio==5.23.3
gradio_client==1.8.0
groovy==0.1.2
grpcio==1.73.1
h11==0.14.0
hjson==3.1.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
hydra-core==1.3.2
idna==3.10
imageio==2.37.0
imgaug==0.4.0
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.5
jedi==0.19.2
Jinja2==3.1.4
jiter==0.9.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
lazy_loader==0.4
llava==0.0.1.dev0
llvmlite==0.44.0
lxml==5.3.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.2.0
multiprocess==0.70.16
nest_asyncio==1.6.0
networkx==3.3
ninja==1.11.1.4
nltk==3.9.1
numba==0.61.2
numpy==1.26.4
nvidia-ml-py==12.570.86
omegaconf==2.3.0
openai==1.70.0
opencv-python==4.11.0.86
orjson==3.10.16
packaging==24.2
pandas==2.2.3
parso==0.8.4
peft==0.15.1
Pillow==9.5.0
pip==25.1
platformdirs==4.3.7
prompt-toolkit==3.0.43
propcache==0.3.0
protobuf==6.30.1
psutil==5.9.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==19.0.1
pycocoevalcap==1.2
pycocotools==2.0.8
pycryptodomex==3.22.0
pydantic==2.11.2
pydantic_core==2.33.1
pydub==0.25.1
Pygments==2.19.1
pynndescent==0.5.13
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytorchyolo==1.7.0
pytz==2025.2
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
ray==2.44.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rouge==1.0.1
rouge_score==0.1.2
rpds-py==0.24.0
ruff==0.11.4
s2wrapper==0.1
safehttpx==0.1.6
safetensors==0.5.3
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
semantic-version==2.10.0
sentencepiece==0.2.0
setproctitle==1.3.5
setuptools==78.1.1
sglang==0.4.5
shapely==2.1.1
shellingham=1.5.4
shortuuid==1.0.13
six==1.17.0
sniffio==1.3.1
stack_data==0.6.3
starlette==0.46.1
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==3.0.1
terminaltables==3.1.10
thop==0.1.1-2209072238
threadpoolctl==3.6.0
tifffile==2025.6.11
tiktoken==0.9.0
tokenizers==0.21.1
tomlkit==0.13.2
torch==2.6.0+cu126
torchsummary==1.5.1
torchtext==0.18.0
torchvision==0.21.0+cu126
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
typer==0.15.2
typing_extensions==4.12.2
typing-inspection==0.4.0
tzdata==2025.2
ultralytics==8.3.160
ultralytics-thop==2.0.14
umap-learn==0.5.7
urllib3==1.26.20
uvicorn==0.34.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.13
xformers==0.0.29.post3
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0